bookstore-micro05/book-service/requirements.txt
Django>=5.0
djangorestframework
mysqlclient
gunicorn
cryptography
tensorflow-cpu>=2.16.1